Tic Tac Toe app first push
App has a Player constructor with 2 New Players instantiated after the
New Game button is pressed.

Board screen and win screen are hidden on page load.

Event handlers are bound to square list items so that the active
player’s SVG animates within a square on hover and the active player’s
symbol remains on click.

On click, the hover event handler is disabled to prevent editing of the
square by another player.

On click, the active player is changed so that a different symbol
appears on hover and is stored in the hovered square on click, another
switch of active player occurs and so on, until the board is full and
all hover binders have been disabled.

Logic is beginning to be written for testing the conditions that create
a win for any active player on click.