Game works for one turn. On the way to requiring implementation of MixMax algorithm.
The game works correctly for one turn; however, the unbinding of the
hover and click effects persists into the second game because the
handlers are still removed.

—————————————

Initialize a player with handlers in the constructor function as a
prototype for the active player?

Beginning of game should create event handler, which is unbound once
for each game played (for each player instantiated).